Former-commit-id: 52fc68275a275b015a40e626d8a41f7a6b4e5411 [formerly 114992cd45848027b676d40a4db634ce874fce6b [formerly 7965f366e5c3b5ff376a9ed914c881daff48a02e]]
Former-commit-id: db60822f20570549c026970a3c1e67e74aff4bb9
Former-commit-id: 03e3ac63284d575d1f3035945489406a08944e27

+++ scripts/msct_shape.py
@@ -324,8 +324,9 @@ def compute_properties_along_centerline(fname_seg_image, property_list, fname_di
         plt.show()
 
     # Removing temporary folder
-    os.chdir('..')
-    shutil.rmtree(path_tmp, ignore_errors=True)
+    if remove_temp_files:
+        os.chdir('..')
+        shutil.rmtree(path_tmp, ignore_errors=True)
 
     return property_list, properties
